requirements.txt
Flask==3.0.3
google-cloud-aiplatform==1.47.0
google-cloud-storage==2.14.0
gunicorn==22.0.0
